利用 clippy 提高代码质量

+++ kern/src/arch/riscv/cpu.rs
@@ -24,8 +24,7 @@ pub fn time() -> Duration {
 
 pub fn set_timer(next: Duration) {
     sbi::timer::set_timer(
-        (next.as_nanos() as u128 * cpus::timebase_freq().unwrap() as u128 / 1_000_000_000u128)
-            as u64,
+        (next.as_nanos() * cpus::timebase_freq().unwrap() as u128 / 1_000_000_000u128) as u64,
     )
     .unwrap();
 }
